requirements.txt
# NOTE: install torch with CUDA >= 12.6 first!
# torch==2.10.0
# torchdata==0.11.0

# main dependencies (must be installed)
torchinfo==1.8.0
pillow==11.0.0
matplotlib==3.10.3
open3d==0.19.0 # https://www.open3d.org/docs/latest/getting_started.html
pyyaml==6.0.2
tqdm==4.67.1
pytransform3d==3.14.1
rich==13.9.4

loguru==0.7.3
opencv-python==4.13.0.90
distance3d==0.9.0
ultralytics